# docs/swagger.yaml
basePath: /api/v7
consumes:
- application/json
definitions:
  models.Bid:
    properties:
      amount:
        type: number
      user:
        $ref: '#/definitions/models.ShortUserData'
    type: object
...
  models.Listing:
    properties:
      active:
        type: boolean
      auctioneer:
        $ref: '#/definitions/models.ShortUserData'
      bids_count:
        type: integer
      category:
        type: string
      closing_date:
        type: string
      desc:
        type: string
      highest_bid:
        type: number
      image:
        type: string
      name:
        type: string
      price:
        type: number
      slug:
        type: string
      time_left_seconds:
        type: integer
      watchlist:
        type: boolean
    type: object
...
  models.ShortUserData:
    properties:
      avatar:
        example: https://my-avatar.com
        type: string
      name:
        example: John Doe
        type: string
    type: object
...
  schemas.AddOrRemoveWatchlistResponseDataSchema:
    properties:
      guestuser_id:
        type: string
    type: object
  schemas.AddOrRemoveWatchlistResponseSchema:
    properties:
      data:
        $ref: '#/definitions/schemas.AddOrRemoveWatchlistResponseDataSchema'
      message:
        example: Data fetched/created/updated/deleted
        type: string
      status:
        example: success
        type: string
    type: object
  schemas.AddOrRemoveWatchlistSchema:
    properties:
      slug:
        example: listing_slug
        type: string
    required:
    - slug
    type: object
  schemas.BidResponseDataSchema:
    properties:
      bids:
        items:
          $ref: '#/definitions/models.Bid'
        type: array
      listing:
        type: string
    type: object
  schemas.BidResponseSchema:
    properties:
      data:
        $ref: '#/definitions/models.Bid'
      message:
        example: Data fetched/created/updated/deleted
        type: string
      status:
        example: success
        type: string
    type: object
  schemas.BidsResponseSchema:
    properties:
      data:
        $ref: '#/definitions/schemas.BidResponseDataSchema'
      message:
        example: Data fetched/created/updated/deleted
        type: string
      status:
        example: success
        type: string
    type: object
...
  schemas.CreateBidSchema:
    properties:
      amount:
        example: 1000
        type: number
    required:
    - amount
    type: object
...
  schemas.ListingDetailResponseDataSchema:
    properties:
      listing:
        $ref: '#/definitions/models.Listing'
      related_listings:
        items:
          $ref: '#/definitions/models.Listing'
        type: array
    type: object
  schemas.ListingDetailResponseSchema:
    properties:
      data:
        $ref: '#/definitions/schemas.ListingDetailResponseDataSchema'
      message:
        example: Data fetched/created/updated/deleted
        type: string
      status:
        example: success
        type: string
    type: object
  schemas.ListingsResponseSchema:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Listing'
        type: array
      message:
        example: Data fetched/created/updated/deleted
        type: string
      status:
        example: success
        type: string
    type: object
...
  utils.ErrorResponse:
    properties:
      data:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
      status:
        type: string
    type: object
...
info:
  contact: {}
  description: A simple bidding API built with Fiber
  title: Bidout Auction API
  version: "7.0"
paths:
...
  /listings:
    get:
      description: This endpoint retrieves all listings.
      parameters:
      - description: Listings Quantity
        in: query
        name: quantity
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.ListingsResponseSchema'
      security:
      - BearerAuth: []
      - GuestUserAuth: []
      summary: Retrieve all listings
      tags:
      - Listings
...
  /listings/categories/{slug}:
    get:
      description: This endpoint retrieves all listings in a particular category.
        Use slug 'other' for category other
      parameters:
      - description: Category Slug
        in: path
        name: slug
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.ListingsResponseSchema'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Retrieve all listings by category
      tags:
      - Listings
  /listings/detail/{slug}:
    get:
      description: This endpoint retrieves detail of a listing.
      parameters:
      - description: Listing Slug
        in: path
        name: slug
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.ListingDetailResponseSchema'
      summary: Retrieve listing's detail
      tags:
      - Listings
  /listings/detail/{slug}/bids:
    get:
      description: This endpoint retrieves at most 3 bids from a particular listing.
      parameters:
      - description: Listing Slug
        in: path
        name: slug
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.BidsResponseSchema'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Retrieve bids in a listing
      tags:
      - Listings
    post:
      description: This endpoint adds a bid to a particular listing.
      parameters:
      - description: Listing Slug
        in: path
        name: slug
        required: true
        type: string
      - description: Create Bid
        in: body
        name: amount
        required: true
        schema:
          $ref: '#/definitions/schemas.CreateBidSchema'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/schemas.BidResponseSchema'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add a bid to a listing
      tags:
      - Listings
  /listings/watchlist:
    get:
      description: This endpoint retrieves all watchlist listings.
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.ListingsResponseSchema'
      security:
      - BearerAuth: []
      - GuestUserAuth: []
      summary: Retrieve all listings by users watchlist
      tags:
      - Listings
    post:
      description: This endpoint adds or removes a listing from a user's watchlist,
        authenticated or not.... As a guest, ensure to store guestuser_id in localstorage
        and keep passing it to header 'guestuserid' in subsequent requests
      parameters:
      - description: Add/Remove Watchlist
        in: body
        name: listing_slug
        required: true
        schema:
          $ref: '#/definitions/schemas.AddOrRemoveWatchlistSchema'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.AddOrRemoveWatchlistResponseSchema'
        "201":
          description: Created
          schema:
            $ref: '#/definitions/schemas.AddOrRemoveWatchlistResponseSchema'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      - GuestUserAuth: []
      summary: Add or Remove listing from a users watchlist
      tags:
      - Listings
produces:
- application/json
securityDefinitions:
  BearerAuth:
    description: Type 'Bearer jwt_string' to correctly set the API Key
    in: header
    name: Authorization
    type: apiKey
  GuestUserAuth:
    description: For guest watchlists. Get ID (uuid) from '/api/v7/listings/watchlist'
      POST endpoint
    in: header
    name: GuestUserId
    type: apiKey
swagger: "2.0"